requirements.txt
discord.py
translate
